FLIZ-60 rename: notificationVariants 이름 변경
