+++ packages/ui/src/components/variants/withNotification.ts
@@ -1,0 +1,7 @@
+export const colorVariants = {
+  default: "bg-background-sub5 text-text-sub5",
+  sub1: "bg-background-sub1 text-text-primary",
+  sub2: "bg-background-sub2 text-text-primary",
+  brand: "bg-brand-primary-500 text-text-primary",
+  destructive: "bg-notification text-text-primary",
+};
